requirements.txt
selenium~=4.1.2
google-api-python-client~=2.39.0
uvicorn~=0.17.5
gunicorn~=20.1.0
fastapi~=0.75.0